Move wayward applyMoves function out of uci. Refactoring and whathaveyous
