+++ source/main.c
@@ -13,7 +13,7 @@ GameState globalGameState;
 int main (int argc, char *argv[]) {
     // Default to start position
     // Different positions will be handled by input parser
-    char cstr[INPUT_SIZE] = DEFAULT_FEN_STRING;
+    char cstr[255] = DEFAULT_FEN_STRING;
 
     if (argc > 1) {
         memcpy(cstr, argv[1], strlen(argv[1]));
